Move recruitUnit code to separate files
