+++ src/vcxproj/ironfist/ironfist.vcxproj.filters
@@ -306,6 +306,9 @@
     <ClCompile Include="..\..\cpp\shared\town\town_globals.cpp">
       <Filter>Source Files\town</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\cpp\shared\gui\recruit_unit.cpp">
+      <Filter>Source Files\gui</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\cpp\shared\stdafx.h">
@@ -520,6 +523,9 @@
     </ClInclude>
     <ClInclude Include="..\..\cpp\shared\town\town_globals.h">
       <Filter>Header Files\town</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\cpp\shared\gui\recruit_unit.h">
+      <Filter>Header Files\gui</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
